requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
python-jobspy==1.1.75
pandas
